Add pseudocode for determining remaining time
Upon the final test I realised the complexity was more than I had anticipated:

- [x] Admin booking 2020-01-01 17:00 -> 2020-01-01 17:00 = TRUE
- [x] Admin booking  2020-01-01 00:00 -> 2020-01-03 = FALSE
- [x] Admin booking 2020-01-01 00:00 -> 2020-01-02 = FALSE
- [x] Admin booking 2020-01-01 00:00 -> 2020-01-01 00:05 = TRUE
- [x] Admin booking 2020-01-01 00:00 -> 2020-01-01 14:00 = TRUE
- [x] Admin booking 2019-10-01 17:00 -> 2020-01-01 24:00 = FALSE
- [x] Admin booking 2019-10-01 17:30 -> 2020-01-01 17:30, Wednesday Regular Hours 09:00->17:00 = FALSE
- [x] Admin booking 2019-10-01 17:30 -> 2020-01-01 17:00, Wednesday Regular Hours 09:00->17:00 = FALSE
- [ ] Admin booking 2019-10-01 17:30 -> 2020-01-01 16:30, Wednesday Regular Hours 09:00->17:00 = TRUE

In the last commit, the code is subtracting admin bookings and regular hours from 24, but this means a time, such as 15:00, is being subtracted "twice". That is, if an admin booking covers 15:00, and a regular hour covers that, the algorithm is removing 15:00 for each type of time.

I am going to handle this by creating an array of "Regular Hours", each of which will have a number representing the time it has left. We will then go through bookings and admin bookings, removing time from that range if the booking is in/affects that range of time