Async Fix: Bookings now update when you create a bookign
Learned:

- Do NOT return a response from the server if you are awaiting on an operation to complete, until the operation has completed
- Define whether you are relying on server response or db operation to complete
- Define whether you are going to use server returned data or db returned data from the server

The issue was that I was awaiting a method, but that method wasn't awaiting for the axios post request.

Front -> Service -> Back-end -> Db

I was getting a response from back-end and then checking db separately before the above chain had even finished doing its db stuff.

So, as a general principle for safety: Do not return a server response until the db update operation is complete - I was returning 202 to say the request was received. But the above timing principle is more useful in itself.