Bug Fix/Optimisation: All available intervals now render
If either an admin booking or customer booking array was null the event wasn't getting added. I was doing it from the POV of customer booking and admin booking separately, whereas this change separates the start and end out to the surround interval loop.

Now:

For each interval .... set interval available

Before:

For each interval ... for admin booking set available ... for customer booking set available.

If that's not enough detail read the code buddy