Express Server Now Handles SPA Routes Navigated to From Browser
If you went to /register you would see "Cannot GET /register" because we are using an SPA, and the routing isn't done from the server, but from Vue. We now direct all requests to the server to /index.html, meaning Vue will do the routing.

The TODO problem is that we won't get a 404 if the user goes to a wrong page. Need to add a 404 page in the Vue application.